% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accrual.R
\docType{class}
\name{AccrualGenerator-class}
\alias{AccrualGenerator-class}
\title{A class to allow additional subjects 
to be recruited onto the study}
\description{
A class to allow additional subjects 
to be recruited onto the study
}
\section{Slots}{

\describe{
\item{\code{f}}{A function with 1 argument (the number of subjects to be recruited)
which returns a vector of recruitment times (of the required length)}

\item{\code{model}}{A string, name of the accrual procedure used}

\item{\code{text}}{The summary text to be output when printing the accrual procedure
for example "a poisson process with rate..."}
}}
\seealso{
\code{\link{show,AccrualGenerator-method}}
}

