% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.R
\docType{class}
\name{AnalysisResults-class}
\alias{AnalysisResults-class}
\title{Class that contains results from the predict function of the \code{Study} object}
\arguments{
\item{critical.HR}{The critical hazard ratio at the time the critical number of events
is expected to occur or NA if single arm trial}

\item{critical.data}{A data frame with 1 row contain the time, event and recruitment details
for the time at which the critical number of events is expected to occur. 
The data frame has 0 rows if the expected time is > study duration or the trial had a single arm}

\item{critical.events.req}{the number of events (not rounded) required for the study's 
requested power and alpha or NA if single arm study}

\item{av.hr}{The average hazard ratio = the study hazard ratio if no lag, see the
vignette for how this is calculated if there is a lag. For single arm studies this is NA.}

\item{grid}{A data frame with rows contain the time, event and recruitment details
at times used to plot the event curves}

\item{predict.data}{A data frame with rows contain the time, event and recruitment details
at times / number of events requested to predicted by the user when calling the predict function.}

\item{study}{The \code{Study} object used to calculate the results}

\item{list}{The \code{Sfn} objects containing the survival functions for the results.}
}
\description{
Three of the slots of this class (grid, critical.data, predict.data) are data frames with
the number of events, recruitment details and cumulative time at risk at given time. See the 
vignette for details of these data frames and the \code{AnalysisResults} class in general.
}

