% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eventData.R, R/singleSimDetails.R
\docType{methods}
\name{CalculateDaysAtRisk}
\alias{CalculateDaysAtRisk}
\alias{CalculateDaysAtRisk,EventData-method}
\alias{CalculateDaysAtRisk,SingleSimDetails-method}
\title{Calculate the number of days at risk}
\usage{
CalculateDaysAtRisk(object, ...)

\S4method{CalculateDaysAtRisk}{EventData}(object)

\S4method{CalculateDaysAtRisk}{SingleSimDetails}(object, times)
}
\arguments{
\item{object}{The object to calculate the days at risk}

\item{...}{Additional arguments passed to the function}

\item{times}{A vector of dates for calculating cutting the simulated date at in order to
calculate the number of days at risk by this point.}
}
\description{
For \code{EventData} object this is the number of days
at risk of the input data. For \code{SingleSimDetails} object
this is the median numberof days at risk in the simulation set at the given time
}

