% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eventData.R
\name{CalculateProgEventTypes}
\alias{CalculateProgEventTypes}
\title{This function calculates the event type for pfs data}
\usage{
CalculateProgEventTypes(has.event, prog.date, dth.date)
}
\arguments{
\item{has.event}{A vector or whether subjects had event}

\item{prog.date}{A vector of dates (or strings in the standard eventPrediction allowed formats)
of progression.date or blank if unknown}

\item{dth.date}{A vector of dates (or strings in the standard eventPrediction allowed formats)
of death date or blank if unknown}
}
\value{
A vector with NA if subject does not have event, 
"Death" if dth.date <= prog.date or prog.date blank,
"Progression (not death)" if prog.date < dth.date or dth.date blank
"Progression (unknown if death)" if both prog.date and dth.date are blank
}
\description{
This function is called before creating the EventData object
}

