% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eventDataDiagnostic.R
\docType{methods}
\name{censorInformation}
\alias{censorInformation}
\alias{censorInformation,EventData-method}
\title{Output a data frame containing all subjects who been censored before
a given date}
\usage{
censorInformation(object, ...)

\S4method{censorInformation}{EventData}(object, censor.date = NULL)
}
\arguments{
\item{object}{An \code{EventData} object}

\item{...}{Additional parameters to be passed to the method}

\item{censor.date}{All subjects who have a censor date before this
date should be output, by default (i.e. when NULL) the latest date for which 
any subject information (withdrawal/event/censor) is known}
}
\value{
A data frame with the required subject's data
}
\description{
Output a data frame containing all subjects who been censored before
a given date not including subjects censored at end of follow up period
}

