% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{csvSniffer}
\alias{csvSniffer}
\title{Function which attempts to read in a csv file
picking a delimiter from a given set of options}
\usage{
csvSniffer(path, delim_options = c(",", ";", "\\t"))
}
\arguments{
\item{path}{The path to the csv file to be read}

\item{delim_options}{A vector of characters from which the
delimiter should be chosen}
}
\value{
If successful a data frame containing the csv file
if unsuccessful an error is thrown
}
\description{
The function works by counting the number of occurences 
of each delimiter on each line and for exactly one 
of the delimiter options all lines contain the same number
of occurences (>0) then this delimiter is used
}

