% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eventData.R
\docType{methods}
\name{CutData}
\alias{CutData}
\alias{CutData,EventData-method}
\title{Cut the Event Data}
\usage{
CutData(object, date)

\S4method{CutData}{EventData}(object, date)
}
\arguments{
\item{object}{The EventData object}

\item{date}{The date to cut the data at}
}
\value{
\code{EventData} object with the data censored at the appropriate point
}
\description{
Create a new EventData from an existing one which shows how the
data would have looked on a given date (assuming no lag in reporting of 
events and if necessary subjects have been censored at cutoff point).
If subject was censored before the cut date then their censored date remains
unchanged
}

