% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eventDataDiagnostic.R
\docType{methods}
\name{DiagnosticPlot}
\alias{DiagnosticPlot}
\alias{DiagnosticPlot,EventData-method}
\title{Output a plot showing the lag between censoring and the date the analysis
is being performed}
\usage{
DiagnosticPlot(object, ...)

\S4method{DiagnosticPlot}{EventData}(object, window.size = NULL,
  analysis.date = NULL, separate.events = TRUE)
}
\arguments{
\item{object}{An \code{EventData} object}

\item{...}{Additional arguments for the function}

\item{window.size}{An optional integer. If used an additional 2 lines at y=x-window.size 
and y = x-2*window.size are drawn on the graph. If \code{window.size} is chosen to be the
visit schedule (in days) then these lines provide an easy way to determine the number of subjects
who have missed one or two visits.}

\item{analysis.date}{The date the analysis is being perfomed on, by default (i.e. when NULL) it is the
the latest date at which any subject is censored/is known to have had an event}

\item{separate.events}{Logical, if FALSE then all events are coloured the same with label "Had Event", if
TRUE then the different event types (object@subject.data$event.type) are coloured individually.}
}
\description{
Output a plot showing the lag between censoring and the date the analysis
is being performed
}

