% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/displayOptions.R
\docType{class}
\name{DisplayOptions-class}
\alias{DisplayOptions-class}
\title{Class defining display options for predict from parameter graphs}
\description{
Class defining display options for predict from parameter graphs
}
\section{Slots}{

\describe{
\item{\code{Time}}{string displaying the time units}

\item{\code{Title}}{string title to be displayed on the graph}

\item{\code{StartDate}}{if `0' the x-axis is labelled using 0,1,2,...
if a date is used then the x-axis is labelled with dates}

\item{\code{Control}}{legend text for control arm}

\item{\code{Exp}}{legend text for experimetal arm}

\item{\code{Trecruit}}{logical in \code{summary} output number recruited}

\item{\code{Tratio}}{logical in \code{summary} output recruitment ratio}

\item{\code{Tacc}}{logical in \code{summary} output recruit details}

\item{\code{Tmedian}}{logical in \code{summary} output control median}

\item{\code{Thr}}{logical in \code{summary} output hazard ratio}

\item{\code{Tcrithr}}{logical in \code{summary} output critical hazard ratio}

\item{\code{text.width}}{numeric in \code{summary} the width for the summary text}

\item{\code{ShowRec}}{logical. Show the number of subjects recruited at target times}

\item{\code{Dropout}}{logical. Show the drop out details}

\item{\code{atRiskTime}}{string displayin the units of at risk,}

\item{\code{atRiskConversion}}{numeric, factor for converting Time into atRiskTime}

\item{\code{showatRisk}}{logical should the at risk details be output}
}}

