% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/displayOptions.R
\name{DisplayOptions}
\alias{DisplayOptions}
\title{DisplayOptions constructor}
\usage{
DisplayOptions(...)
}
\arguments{
\item{...}{parameters passed to new}
}
\value{
A new DisplayOptions object
}
\description{
DisplayOptions constructor
}
\seealso{
\code{\link{DisplayOptions-class}}
}

