% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eventData.R
\name{EmptyEventData}
\alias{EmptyEventData}
\title{Create an \code{EventData} object with no subjects}
\usage{
EmptyEventData(followup = Inf)
}
\arguments{
\item{followup}{The time subjects are followed until censoring}
}
\description{
Create an \code{EventData} object with no subjects
}

