% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eventPrediction_package.R
\docType{data}
\name{event.data}
\alias{event.data}
\title{Simulated data that was originally obtained by the exportData function available
in the package. The randomization dates have not been modelled properly and is only
evenly distributed in the study period.}
\format{A data.frame containing a row per patient with six columns:
subject patient identifier
has.event Logical indicating whether patient has had an event or not
event.date Date of event
last.date  The date of the last visit
rand.date Date when patient was randomized. 
eventType For subjects with an event, the type of event which occurred}
\description{
Simulated data that was originally obtained by the exportData function available
in the package. The randomization dates have not been modelled properly and is only
evenly distributed in the study period.
}

