% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eventData.R
\docType{class}
\name{EventData-class}
\alias{EventData-class}
\title{Class representing data to use for new predictions}
\description{
Class representing data to use for new predictions
}
\section{Slots}{

\describe{
\item{\code{subject.data}}{a data frame with 6 columns
"subject", "rand.date", "has.event", "withdrawn", "censored.at.follow.up" and "time" and "site" and "event.type"
see vignette and \code{EventData} for further details.}

\item{\code{followup}}{A numeric value for the fixed follow up period a subject
is followed. This is in days. If there is no fixed followup period then 
Inf should be used}
}}

