% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eventModel.R
\docType{class}
\name{EventModel-class}
\alias{EventModel-class}
\title{An S4 class containing a fitted survival model
of an Eventdata object}
\description{
An S4 class containing a fitted survival model
of an Eventdata object
}
\section{Slots}{

\describe{
\item{\code{model}}{An S3 survreg object}

\item{\code{event.data}}{An EventData object used to fit a Weibull survial model}

\item{\code{simParams}}{The \code{FromDataSimParam} object which contains the information
needed to generate the survial times}
}}

