% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eventPrediction_package.R
\docType{package}
\name{eventPrediction-package}
\alias{eventPrediction}
\alias{eventPrediction-package}
\title{Expected Progress of a Survival Study Assuming Exponential/Weibull Survival}
\description{
Expected Progress of a Survival Study
}
\details{
Calculate the required number of events for a two-arm survival
study, and the expected accumulation of events assuming exponential
or Weibull survival. Non-uniform accrual may be specified. It is also
possible to use real data for predicting the total number of events at a
future time point.
}
\author{
\email{daniel.dalevi@astrazeneca.com}
}

