% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eventData.R
\docType{methods}
\name{EventVsTimePlot}
\alias{EventVsTimePlot}
\alias{EventVsTimePlot,EventData-method}
\title{Makes a barplot with the number of events over time}
\usage{
EventVsTimePlot(object, timeunit = "Months", ...)

\S4method{EventVsTimePlot}{EventData}(object, timeunit = "month")
}
\arguments{
\item{object}{The object to create the event versus time plot}

\item{timeunit}{The resoloution on the x-axis (month, weeks or quarter)}

\item{...}{Additional arguments passed to the function}
}
\description{
Makes a barplot with the number of events over time
}

