% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eventData.R
\docType{methods}
\name{fit}
\alias{fit}
\alias{fit,EventData-method}
\title{Fit a survival model to an EventData object}
\usage{
\S4method{fit}{EventData}(object, dist = "weibull")
}
\arguments{
\item{object}{an EventData object}

\item{dist}{character, either weibull or loglogistic, which model
is to be fit}

\item{...}{Additional arguments to the function}
}
\value{
an EventModel object
}
\description{
Fit a survival model to an EventData object
}

