% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{FitMixtureModel}
\alias{FitMixtureModel}
\title{Given trial assumptions describing an exponential model
Use the method of moments to fit a single Weibull distribution}
\usage{
FitMixtureModel(HR, r, M)
}
\arguments{
\item{HR}{Hazard ratio}

\item{r}{Randomization balance}

\item{M}{control arm median}
}
\description{
See, for example, Lei, Y. Evaluation of three methods for estimating the Weibull distribution 
parameters of Chinese pine (Pinus tabulaeformis). Journal of Forest Science 54.12 (2008): 566-571.
}

