% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fromDataSimParam.R
\docType{methods}
\name{FromDataParam}
\alias{FromDataParam}
\alias{FromDataParam,missing,character-method}
\alias{FromDataParam,survreg,character-method}
\title{Method to create FromDataSimParam objects}
\usage{
FromDataParam(object, type, ...)

\S4method{FromDataParam}{survreg,character}(object, type)

\S4method{FromDataParam}{missing,character}(object, type, rate, shape,
  sigma = matrix(c(0, 0, 0, 0), nrow = 2))
}
\arguments{
\item{object}{The object to derive the fromDataParam}

\item{type}{The type argument to be used in the FromDataSimParam type slot}

\item{...}{Additional options to be passed into the function}

\item{rate}{The (mean) Weibull/loglogistic rate parameter for the simulations}

\item{shape}{The (mean) Weibull/loglogistic shape parameter for the simulations}

\item{sigma}{Advanced: The covariance (uncertainty) matrix used to sample single
simulation rates and shapes. The covariance matrix describes the uncertainty of
{-log(rate),-log(shape)} - i.e. the var matrix of a survreg Weibull/loglogistic survival model}
}
\description{
Method to create FromDataSimParam objects

Create FromDataSimParam object from survreg model
}

