% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataResults.R
\docType{class}
\name{FromDataResults-class}
\alias{FromDataResults-class}
\title{Class that contains results from the simulating event predictions
from Data}
\description{
Class that contains results from the simulating event predictions
from Data
}
\section{Slots}{

\describe{
\item{\code{limit}}{The confidence interval width used by recQuantiles, eventQuantiles and dropoutQuantiles
If limit = 0.05 then the CI is the 5th - 95th percentile}

\item{\code{eventQuantiles}}{The dates on which the median and CI of each event is expected to occur}

\item{\code{event.data}}{The EventData object used when simulating events}

\item{\code{accrualGenerator}}{The AccrualGenerator object used to recruit new subjects}

\item{\code{Naccrual}}{The number of additional subjects recruited for each simulation}

\item{\code{time.pred.data}}{A data frame giving the expected number of events (and CI intervals) for
given target dates. See  vignette for further details}

\item{\code{event.pred.data}}{A data frame giving the expected date (and CI intervals) for given 
target numbers of events occurring. See vignette for further details}

\item{\code{recQuantiles}}{The dates on which the median and CI of each recruitment is expected to occur}

\item{\code{dropoutQuantiles}}{The dates on which the median and CI of each event is expected to occur}

\item{\code{singleSimDetails}}{A SingleSimDetails object containing the subject level details of the simulation
see SingleSimDetails class documentaion for further details.}

\item{\code{dropout.shape}}{The Weibull shape parameter of the dropout risk used for the simulations}

\item{\code{dropout.rate}}{The Weibull rate parameter of the dropout risk used for the simulations (in units of day^{-1}).
If no dropout then this is 0}

\item{\code{simParams}}{The \code{FromDataSimParam} object which contains the information
used to generate the survial times}
}}
\seealso{
\code{\link{show,FromDataResults-method}}, \code{\link{predict,FromDataResults-method}},
\code{\link{summary,FromDataResults-method}}
}

