% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fromDataSimParam.R
\docType{class}
\name{FromDataSimParam-class}
\alias{FromDataSimParam-class}
\alias{FromDataSimParam.class}
\title{An S4 class containing the function required
to simulate conditional survival times}
\description{
An S4 class containing the function required
to simulate conditional survival times
}
\section{Slots}{

\describe{
\item{\code{type}}{A character string describing the distribution used
for example "Weibull"}

\item{\code{generateParameterFunction}}{A function with a single parameter Nsim which returns
a matrix with Nsim rows and columns Id(=1:Nsim) and the specific parameters needed by
conditionalFunction below. In the Weibull case these are the specific rate and shape 
parameters used for each simulation}

\item{\code{conditionalFunction}}{A function with the following arguments t.conditional, params, HR.
t.conditional is a vector of current survival times, params is a row of the matrix returned
by generateParameterFunction and HR is a vector of 1s for subjects in control arm and the hazard ratio 
for subjects in the active arm (if HR argument to simulate is NULL then this is a vector of 1s). The function
then returns a vector of event times.}

\item{\code{parameters}}{A list of parameters used for the simulation. For the Weibull this is
rate, shape and sigma (the uncertainty matrix of the parameters - specifically the covariance matrix from survreg)}
}}

