% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accrual.R
\name{Generate.PoissonAccrual}
\alias{Generate.PoissonAccrual}
\title{Create A Poisson process accrual generator}
\usage{
Generate.PoissonAccrual(start.date, rate)
}
\arguments{
\item{start.date}{The start date for the recruitment process}

\item{rate}{The rate of the required Poisson process.}
}
\value{
An \code{AccrualGenerator} object which can 
be used when recruiting subjects with a Possion process
with given \code{start.date} and \code{rate}
}
\description{
Create A Poisson process accrual generator
}

