% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lag.R
\docType{methods}
\name{isNullLag}
\alias{isNullLag}
\alias{isNullLag,LagEffect-method}
\title{Function to check whether Lag is a null lag}
\usage{
isNullLag(Lag)

\S4method{isNullLag}{LagEffect}(Lag)
}
\arguments{
\item{Lag}{A LagEffect object}
}
\value{
TRUE if null lag, FALSE otherwise
}
\description{
This checks whether the Lag.T slot is 0
}

