% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study.R
\docType{methods}
\name{isSingleArm}
\alias{isSingleArm}
\alias{isSingleArm,Study-method}
\title{Is the \code{Study} a single arm study}
\usage{
isSingleArm(study)

\S4method{isSingleArm}{Study}(study)
}
\arguments{
\item{study}{A \code{Study} object}
}
\value{
TRUE if study is single arm, FALSE otherwise
}
\description{
Is the \code{Study} a single arm study
}

