% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lag.R
\docType{class}
\name{LagEffect-class}
\alias{LagEffect-class}
\title{Parameter settings for when including a lagged effect.}
\description{
Note the lambda and HR in the \code{Study} class will be used
for time > T
}
\section{Slots}{

\describe{
\item{\code{Lag.T}}{Lagtime (T)}

\item{\code{ctrlSpec}}{Control median specification for time period [0,T]}

\item{\code{L.HazardRatio}}{Hazard for time period [0,T]}
}}
\seealso{
\code{\link{show,LagEffect-method}}
}

