% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lag.R
\name{LagEffect}
\alias{LagEffect}
\title{LagEffect constructor}
\usage{
LagEffect(Lag.T, L.Ctr.median = as.numeric(NA),
  L.HazardRatio = as.numeric(NA))
}
\arguments{
\item{Lag.T}{Lagtime (T)}

\item{L.Ctr.median}{Control median for time period [0,T]}

\item{L.HazardRatio}{Hazard for time period [0,T]}
}
\value{
A LagEffect object
}
\description{
Note the lambda and HR in the study class will be used
for time > T
}

