% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfn.R, R/results.R
\docType{methods}
\name{LatexSurvivalFn}
\alias{LatexSurvivalFn}
\alias{LatexSurvivalFn,AnalysisResults-method}
\alias{LatexSurvivalFn,Sfn-method}
\title{Method to output a Latex String of the object}
\usage{
LatexSurvivalFn(results, ...)

\S4method{LatexSurvivalFn}{Sfn}(results, decimalplaces, lambda, shape)

\S4method{LatexSurvivalFn}{AnalysisResults}(results, decimalplaces = 3)
}
\arguments{
\item{results}{Object to output}

\item{decimalplaces}{The number of decimal places to output for the 
rate and shape parameters}

\item{lambda}{The symbol to be used for the arm's rate 
parameter, if lagged study then this is a vector of before and after lag
symbols. The latex backslash character needs to be escaped}

\item{shape}{The symbol to be used for Weibull shape parameter}

\item{\ldots}{Additional parameters to be passed into the function}
}
\value{
A latex string of the object (for \code{Sfn} this is the Survival function)
}
\description{
Method to output a Latex String of the object
}

