% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/longlagSettings.R
\docType{class}
\name{LongLagSettings-class}
\alias{LongLagSettings-class}
\title{Class which determines how the simulation (from data) 
procedure deals with subjects
who are censored a long time before the analysis date.}
\description{
Class which determines how the simulation (from data) 
procedure deals with subjects
who are censored a long time before the analysis date.
}
\section{Slots}{

\describe{
\item{\code{analysis.date}}{The date the event prediction is performed on. If \code{as.Date(NA)}
then the latest censor/event date from the data set will be used as the analysis.date}

\item{\code{ndays}}{All subjects who were censored more than \code{ndays} before
\code{analysis.date} will be changed}

\item{\code{toWithdraw}}{Logical. If TRUE then the selected subjects will be withdrawn
rather than be censored. If FALSE, their censor date is replaced by \code{analysis.date}}

\item{\code{visitschedule}}{number of days between visits for pfs data. If non-zero and 
\code{toWithdraw} is TRUE then the selectedsubjects are censored on the date their
latest visit would have been. See vignette for further details}

\item{\code{text}}{The text to be disaplyed when printing the LongLagSettings}
}}
\seealso{
\code{\link{show,LongLagSettings-method}}
}

