% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/longlagSettings.R
\name{LongLagSettings}
\alias{LongLagSettings}
\title{Constructor for \code{LongLagSettings} object}
\usage{
LongLagSettings(analysis.date = as.Date(NA), ndays, toWithdraw,
  visitschedule = 0)
}
\arguments{
\item{analysis.date}{The date the event prediction is performed on. If \code{as.Date(NA)}
then the latest censor/event date from the data set will be used as the analysis.date}

\item{ndays}{All subjects who were censored more than \code{ndays} before
\code{analysis.date} will be changed}

\item{toWithdraw}{Logical. If TRUE then the selected subjects will be withdrawn
rather than be censored. If FALSE, their censor date is replaced by \code{analysis.date}}

\item{visitschedule}{number of days between visits for pfs data. If non-zero and 
\code{toWithdraw} is TRUE then the selectedsubjects are censored on the date their
latest visit would have been. See vignette for further details}
}
\value{
A \code{LongLagSettings} object
}
\description{
Constructor for \code{LongLagSettings} object
}

