% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lag.R
\name{NullLag}
\alias{NullLag}
\title{Create a LagEffect object with no lag}
\usage{
NullLag()
}
\value{
A LagEffect object for which \code{isNullLag()} is TRUE
}
\description{
Create a LagEffect object with no lag
}

