% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eventData.R
\docType{methods}
\name{OnlyUseRecTimes}
\alias{OnlyUseRecTimes}
\alias{OnlyUseRecTimes,EventData-method}
\title{Method to create new \code{EventData} object only including recruitment times}
\usage{
OnlyUseRecTimes(object)

\S4method{OnlyUseRecTimes}{EventData}(object)
}
\arguments{
\item{object}{The  object from which to create the new \code{EventData} object}
}
\value{
An \code{EventData} object with time=0 and no events or withdrawn subjects
}
\description{
Method to create new \code{EventData} object only including recruitment times
}

