% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eventPrediction_package.R, R/dataResults.R, R/study.R
\docType{methods}
\name{predict}
\alias{predict}
\alias{predict,FromDataResults-method}
\alias{predict,Study-method}
\title{The predict methods for EventPrediction package}
\usage{
\S4method{predict}{FromDataResults}(object, time.pred = NULL,
  event.pred = NULL)

\S4method{predict}{Study}(object, time.pred = NULL, event.pred = NULL,
  step.size = 0.5)
}
\arguments{
\item{object}{The object for which the prediction is required}

\item{time.pred}{If \code{object} is a \code{FromDataResults} object then: 
The Dates for which the number of events
are required at. This should be a vector of character strings,
e.g. c("2017-10-10","2018-04-27) or NULL. 
If it is a \code{Study} object then this is a numeric vector of 
dates (in months from the start of the trial) for which the number
of events are required at.}

\item{event.pred}{The target number of events for which the expect time
is to be predicted, should be a numeric vector}

\item{step.size}{The resolution of the grid to be used to calculate the time v 
expected events curves}

\item{...}{Additional objects}
}
\value{
The predicted values. If \code{object} is \code{Study} the result will
be an \code{AnalysisResults} object.
If \code{object} is a \code{FromDataResults} object it will be
returned with \code{event.pred.data} and/or \code{time.pred.data} slots updated
}
\description{
The predict methods for EventPrediction package
}
\seealso{
\code{\link{FromDataResults-class}} \code{\link{AnalysisResults-class}}
}

