% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eventModel.R
\docType{methods}
\name{risk.table}
\alias{risk.table}
\alias{risk.table,EventModel-method}
\title{Calculate KM risk.table}
\usage{
risk.table(x, ...)

\S4method{risk.table}{EventModel}(x, units = "Days", risk.times = NULL, ...)
}
\arguments{
\item{x}{An \code{EventModel} object}

\item{...}{Additional arguments to pass to survfit}

\item{units}{Scale for the risk table: "Days", "Months" or "Years"}

\item{risk.times}{The times for the risk table (in \code{units}). 
If NULL then the 0, 0.25,0.5,0.75,1 quantiles of the KM survival table are used}
}
\value{
A data frame contianing the number at risk at the given times
}
\description{
Calculate KM risk.table
}

