% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfn.R
\docType{class}
\name{Sfn-class}
\alias{Sfn-class}
\title{A Class containing the Survival Function (for a single arm) 
used in the integral to calculate event times
in predict from parameters. It should not be created by the user
but is created for the user (see sfns slot of \code{AnalysisResult})}
\description{
A Class containing the Survival Function (for a single arm) 
used in the integral to calculate event times
in predict from parameters. It should not be created by the user
but is created for the user (see sfns slot of \code{AnalysisResult})
}
\section{Slots}{

\describe{
\item{\code{sfn}}{A function sfn(x) is the function to be included in the events.integrate 
procedure when using predict from parameters. 
Specifically sfn(x) =  1 - P(had event [i.e. not dropout] by time x) and when no drop outs 
this is exactly the survival function.
If using dropouts this is not quite the survival function
However, the output of LatexSurvivalFn(object) is the survival function.}

\item{\code{SurvivalFunction}}{The actual survival function (if drop outs/finite followup are used then this will
not equal sfn). If finite follow up is used then S(x) = 0 for x > followup}

\item{\code{pdf}}{The pdf function associated with the survival function. If finite follow up is used
then pdf(x) should  = 0 for x > followup}

\item{\code{nullf}}{Logical, TRUE if the object represents NULL (i.e. a survival function for 
a second arm in a single arm study)}

\item{\code{lambda}}{The rate parameters for the arm of trial.
In a trial with lag, this is the rate parameters for time > T}

\item{\code{lambdaot}}{If a lag was used then the rate parameters 
for time < T otherwise NA}

\item{\code{shape}}{The Weibull shape parameter}

\item{\code{LagT}}{The lagtime for the survival function (0 for no lag)}

\item{\code{followup}}{The follow up time for each subject 
(Inf for studies with no fixed followup)}

\item{\code{dropout.shape}}{The Weibull shape parameter for the drop out hazard function}

\item{\code{dropout.lambda}}{The rate parameter for the drop out hazard function = 0 if no dropout}
}}

