% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simQOutput.R, R/eventPrediction_package.R, R/eventData.R, R/accrual.R, R/longlagSettings.R, R/dataResults.R, R/eventModel.R, R/lag.R, R/study.R, R/results.R
\docType{methods}
\name{show}
\alias{show}
\alias{show,AccrualGenerator-method}
\alias{show,AnalysisResults-method}
\alias{show,EventData-method}
\alias{show,EventModel-method}
\alias{show,FromDataResults-method}
\alias{show,LagEffect-method}
\alias{show,LongLagSettings-method}
\alias{show,SimQOutput-method}
\alias{show,Study-method}
\title{The S4 show methods for EventPrediction package}
\usage{
\S4method{show}{SimQOutput}(object)

\S4method{show}{EventData}(object)

\S4method{show}{AccrualGenerator}(object)

\S4method{show}{LongLagSettings}(object)

\S4method{show}{FromDataResults}(object)

\S4method{show}{EventModel}(object)

\S4method{show}{LagEffect}(object)

\S4method{show}{Study}(object)

\S4method{show}{AnalysisResults}(object)
}
\arguments{
\item{object}{The object to be shown}
}
\description{
The S4 show methods for EventPrediction package
}

