% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simQOutput.R
\docType{class}
\name{SimQOutput-class}
\alias{SimQOutput-class}
\title{A class containing the expected (and CI) dates for a specific item of interest}
\description{
For example subject recruitment, events occurring, or subject dropout. 
Each item of interest will have its own SimQOutput object. This class will
be created for the user and does not need to be manually created
}
\details{
As an example median[1] is the expected date of the first specific item of interest
whereas upper[1] is the upper CI of the first specific item of interest.
}
\section{Slots}{

\describe{
\item{\code{median}}{A vector of the expected (i.e median) dates 
of the first, second, ..., item of interest
The length of the vector is the total number of this type of item.}

\item{\code{upper}}{A vector of the expected dates for the upper CI 
of the first, second, ..., item of interest}

\item{\code{lower}}{A vector of the expected dates for the lower CI
of the first, second, ..., item of interest}
}}
\seealso{
\code{\link{simulate,EventModel,missing,missing-method}} \code{\link{FromDataResults-class}}
}

