% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study_constructors.R
\name{SingleArmStudy}
\alias{SingleArmStudy}
\title{Create a \code{Study} object for a single arm trial}
\usage{
SingleArmStudy(N, study.duration, ctrl.median, k, acc.period, shape = 1,
  dropout = NULL, lag.settings = NullLag())
}
\arguments{
\item{N}{Number of subjects to be recruit (integer)}

\item{study.duration}{Number of months the study will be going.}

\item{ctrl.median}{Median time to an event in the control group.}

\item{k}{non-uniformity of accrual (integer, 1=uniform). Non-uniform accrual is allowed for 
using the following distribution for the probability of a patient entering the trial at time \eqn{b} 
within the accrual period \eqn{[0,B]}: \eqn{F(b)=b_k/B_k}; \eqn{f(b)=k b_{k-1}/B_k} where \eqn{k} is the 
measure of non-uniformity (\eqn{k>0}). \eqn{k=1} indicates uniform accrual. This implies that during 
the first half of the accrual period, \eqn{1/2^k} of the patients will be recruited. Half of the patients 
will be recruited by time \eqn{B/2^{1/k}}.}

\item{acc.period}{Accrual time.}

\item{shape}{The Weibull shape parameter}

\item{dropout}{if subjects drop out in study (due to competing risks not as there is a finite follow up time)
then this argument should contain a list with proportion and time and optionally shape i.e.
\code{dropout=list(proportion=0.03,time=12,shape=1.2)} meaning in the absence of events 3% of subjects
will have dropped out after 12 months with a Weibull hazard rate with shape=1.2. If shape is not included then 
it defaults to 1 (exponential rate). If dropout is NULL then no subjects will drop out}

\item{lag.settings}{The \code{LaggedEffect} object describing any lag effect for the study}
}
\value{
A \code{Study} object
}
\description{
Create a \code{Study} object for a single arm trial
}

