% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singleSimDetails.R
\docType{class}
\name{SingleSimDetails-class}
\alias{SingleSimDetails-class}
\title{Class containing the subject level event data
for each simulation}
\description{
A class which contains the recruitment times
event times and event types of all subjects and all
simulations for a predict from data simulation. 
This class is created when performing simulation and does not 
need to be constructed manually
}
\section{Slots}{

\describe{
\item{\code{event.type}}{A matrix (1 row per subject, 1 column per simulation) of
the reason subjects leave trial. 0 is have event, 1 is drop out and 2 is drop out due to follow up}

\item{\code{event.times}}{A matrix (1 row per subject, 1 column per simulation) of
dates (as numeric) subjects leave the trial}

\item{\code{rec.times}}{A matrix (1 row per subject, 1 column per simulation) of
recruitment dates (as numeric). 
Therefore event.times[i,j] - rec.times[i,j] + 1 is the time on study for
subject i in simulation j}
}}

