% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{standarddaysinyear}
\alias{standarddaysinyear}
\title{Return the number of days in the year}
\usage{
standarddaysinyear()
}
\value{
The number of days in the year
}
\description{
This allows the user control over
whether the package uses 365.25, 365 (or any other number) of days
in a year.
}
\details{
The number of days in a month is then set as daysinyear()/12
}

