% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study.R
\docType{class}
\name{Study-class}
\alias{Study-class}
\title{Class defining the Study}
\description{
Class defining the Study
}
\section{Slots}{

\describe{
\item{\code{HR}}{Hazard ratio to be detected}

\item{\code{alpha}}{Significance level [0,1] (see also two-sided indicator)}

\item{\code{power}}{Power [0,1]}

\item{\code{two.sided}}{If TRUE, two sided test will be used (i.e. alpha/2).}

\item{\code{r}}{Control:Experimental subject balance (1:r), i.e. nE/nC=r. r=1 corresponds to equally 
many subjects in both arms. 2 means we have twice the number of subjects in the experimental arm.
Specifically \code{floor(r*N/(r+1))} subjects are 
allocated to the experimental arm and all other subjects are allocated to the control arm.}

\item{\code{N}}{Number of subjects to be recruited (integer)}

\item{\code{study.duration}}{Number of months the study will be going.}

\item{\code{ctrlSpec}}{A CtrlSpec object which calculates the control group median. This object will be created automatically
when calling a constructor for the Study class.}

\item{\code{dropout}}{A list of CtrlSpec object which calculates the median drop out rate for the control arm (index 1) and 
active arm (index 2). 
This object will be created automatically when calling a constructor for the study class}

\item{\code{dropout.shape}}{The Weibull shape parameter of the dropout hazard function}

\item{\code{k}}{non-uniformity of accrual (integer, 1=uniform). Non-uniform accrual is allowed for 
using the following distribution for the probability of a patient entering the trial at time \eqn{b} 
within the accrual period \eqn{[0,B]}: \eqn{F(b)=b_k/B_k}; \eqn{f(b)=k b_{k-1}/B_k} where \eqn{k} is the 
measure of non-uniformity (\eqn{k>0}). \eqn{k=1} indicates uniform accrual. This implies that during 
the first half of the accrual period, \eqn{1/2^k} of the patients will be recruited. Half of the patients 
will be recruited by time \eqn{B/2^{1/k}}.}

\item{\code{acc.period}}{Accrual time in months}

\item{\code{shape}}{The Weibull shape parameter}

\item{\code{followup}}{The time a subject is followed after randomization, if Inf then there is no fixed time period}

\item{\code{type}}{Character: The study type, either "Oncology" or "CRGI"}

\item{\code{lag.settings}}{The \code{LaggedEffect} object describing any lag effect for the study}
}}

