% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eventPrediction_package.R, R/eventData.R, R/dataResults.R, R/results.R
\docType{methods}
\name{summary}
\alias{summary}
\alias{summary,AnalysisResults-method}
\alias{summary,EventData-method}
\alias{summary,FromDataResults-method}
\title{The S4 summary methods for EventPrediction package}
\usage{
\S4method{summary}{EventData}(object)

\S4method{summary}{FromDataResults}(object, round.method = "None",
  text.width = 60, show.predictions = TRUE, show.at.risk = TRUE)

\S4method{summary}{AnalysisResults}(object, options = DisplayOptions())
}
\arguments{
\item{object}{The object for which a summary is desired}

\item{round.method}{If the string "toMonths" then dates are rounded 
to the nearest month. For the lower confidence interval value the month of 
the date 15 days earlier than the given value is used
and for the upper confidence interval value the month of the
date 15 days later is used. For the median, the month is used.}

\item{text.width}{The width of the summary text}

\item{show.predictions}{Logical if TRUE then include the time.pred.data and
event.pred.data information in the text}

\item{show.at.risk}{Output the median number of at risk years}

\item{options}{A \code{DisplayOptions} object}
}
\description{
The S4 summary methods for EventPrediction package
}

